backend/requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.0b0; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.6.0; python_version >= '3.8'
anyio==3.7.1; python_version >= '3.7'
async-timeout==4.0.3; python_version < '3.11'
attrs==23.1.0; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
exceptiongroup==1.1.3; python_version < '3.11'
fastapi==0.104.0
frozenlist==1.4.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httptools==0.6.1
idna==3.4; python_version >= '3.5'
llm-stack-kit==0.1.5
multidict==6.0.4; python_version >= '3.7'
openai==0.28.1
pydantic==2.4.2
pydantic-core==2.10.1; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
pyyaml==6.0.1
requests==2.31.0; python_version >= '3.7'
sniffio==1.3.0; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
tqdm==4.66.1; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==2.0.7; python_version >= '3.7'
uvicorn[standard]==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
websockets==11.0.3
yarl==1.9.2; python_version >= '3.7'
